Struct numrs::matrix::Matrix

pub struct Matrix<T: Number> {
    // some fields omitted
}

2D Matrix

Methods

impl<T: Number> Matrix<T>

fn num_rows(&self) -> usize

Returns the number of rows in the matrix.

fn num_cols(&self) -> usize

Returns the number of columns in the matrix.

fn get(&self, i: usize, j: usize) -> T

Returns the element at ith row and jth column of the matrix. The row, column indexing is 0-based.

fn set(&mut self, i: usize, j: usize, num: T)

Sets the element at ith row and jth column of the matrix as num. The row, column indexing is 0-based.

fn new(rows: usize, cols: usize, default: T) -> Matrix<T>

Creates a new Matrix with dimensions as rows x cols with all values instantiated to default.

fn reshape(&mut self, newrows: usize, newcols: usize)

Resizes the dimensions of the matrix with the new dimensions as newrows x newcols.

fn get_vec(&self) -> Vec<T>

Returns the matrix as a cloned vector in row major order.

fn transpose(&mut self)

Transposes the matrix.

fn trace(&self) -> T

Trace of the matrix.